+++ General/General.vb
@@ -38,15 +38,6 @@ Public Class Folder
         End Get
     End Property
 
-    Shared Property Current As String
-        Get
-            Return Environment.CurrentDirectory.FixDir
-        End Get
-        Set(value As String)
-            Environment.CurrentDirectory = value
-        End Set
-    End Property
-
     Shared ReadOnly Property Temp As String
         Get
             Return Path.GetTempPath.FixDir