+++ Changelog.md
@@ -8,7 +8,10 @@
 - Advanced Info feature of Code Editor has new VapourSynth ClipInfo option. (stax76)
 - Fix items can't be removed from or rearranged in lists (Dendraspis, [#433](/../../issues/433))
 - Fix randomly appearing IOException when using x265 with chunks (Dendraspis, [#431](/../../issues/431))
-
+- Revert current folder of process being set to video temp folder because
+  it locks the folder and also because the .NET Framework allows only
+  260 characters for the working directory of a process. (stax76)
+  
 
 2.1.7.0 Stable (2021-01-07)
 ===========================